chore\clear keyboard autofocus in log in
