+++ app/src/main/java/com/ucm/informatica/spread/Utils/SectionsPagerAdapter.java
@@ -1,0 +1,23 @@
+package com.ucm.informatica.spread.Utils;
+
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentPagerAdapter;
+
+public class SectionsPagerAdapter extends FragmentPagerAdapter {
+
+    public SectionsPagerAdapter(FragmentManager fm) {
+        super(fm);
+    }
+
+    @Override
+    public Fragment getItem(int position) {
+
+        return PlaceholderFragment.newInstance(position);
+    }
+
+    @Override
+    public int getCount() {
+        return 5;
+    }
+}
