+++ app/src/main/res/layout/activity_sign_in.xml
@@ -24,7 +24,7 @@
         app:animationDuration="2500"
         app:layout_constraintBottom_toTopOf="@+id/imageView"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.65"
+        app:layout_constraintHorizontal_bias="0.55"
         app:layout_constraintStart_toStartOf="parent" />
 
 
@@ -49,7 +49,7 @@
         android:id="@+id/subtitleSignIn"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
+        android:layout_marginTop="@dimen/smallMargin"
         android:layout_marginHorizontal="@dimen/largeMargin"
         android:text="Introduce una contraseña para crear tu cuenta"
         android:textColor="@color/white"
@@ -63,7 +63,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginHorizontal="@dimen/mediumMargin"
-        android:layout_marginTop="8dp"
+        android:layout_marginTop="@dimen/smallMargin"
         android:background="@drawable/round_button_shape_dark"
         android:backgroundTint="@color/white"
         android:orientation="horizontal"
